requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
flake8==7.1.1
black==24.8.0
isort==5.13.2
pre-commit==3.8.0
python-dotenv==1.0.1
databases==0.9.0
alembic==1.13.2
SQLAlchemy==2.0.32
PyMySQL==1.1.1
pydantic-settings==2.4.0
aiomysql==0.2.0
cryptography==43.0.0
